Dev 1/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.2
arabic-reshaper==2.1.3
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cramjam==2.5.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.4
et-xmlfile==1.1.0
executing==0.8.2
fake-useragent==0.1.11
fastparquet==0.8.0
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.0
ipython==8.0.1
jedi==0.18.1
kiwisolver==1.3.2
lxml==4.7.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.1
parse==1.19.0
parso==0.8.3
pathspec==0.9.0
pdf2image==1.16.0
persiantools==2.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycares==4.1.2
pycparser==2.21
pyee==8.2.2
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyppeteer==1.0.2
pyquery==1.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.1.4
tomli==2.0.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.1
wsproto==1.0.0
xmltodict==0.12.0
yarl==1.7.2
zipp==3.7.0
